Move aggro and damage modules
