+++ combat/engine/__init__.py
@@ -2,8 +2,8 @@
 
 from ..combatants import _current_hp, CombatParticipant
 from .turn_manager import TurnManager
-from .aggro_tracker import AggroTracker
-from .damage_processor import DamageProcessor
+from ..aggro_tracker import AggroTracker
+from ..damage_processor import DamageProcessor
 from .combat_math import CombatMath
 from .combat_engine import CombatEngine
