+++ world/tests/test_aggro_tracker.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import MagicMock, patch
 
-from combat.engine.aggro_tracker import AggroTracker
+from combat.aggro_tracker import AggroTracker
 
 
 class TestAggroTracker(unittest.TestCase):
@@ -11,9 +11,9 @@ class TestAggroTracker(unittest.TestCase):
         victim = MagicMock(pk=2)
         active = [attacker]
 
-        with patch("combat.engine.aggro_tracker.state_manager.get_effective_stat", return_value=0), \
-             patch("combat.engine.aggro_tracker.state_manager.calculate_xp_reward", return_value=5), \
-             patch("combat.engine.aggro_tracker.award_xp"):
+        with patch("combat.aggro_tracker.state_manager.get_effective_stat", return_value=0), \
+             patch("combat.aggro_tracker.state_manager.calculate_xp_reward", return_value=5), \
+             patch("combat.aggro_tracker.award_xp"):
             tracker.track(victim, attacker)
             assert victim in tracker.table
             tracker.award_experience(attacker, victim, active)
